Replaced size_t whenever possible

+++ input-parsing/input_file.cc
@@ -8,12 +8,12 @@
 static const std::string WHITESPACE = " \n\r\t\f\v";
 std::string ltrim(const std::string& s)
 {
-	size_t start = s.find_first_not_of(WHITESPACE);
+	int start = s.find_first_not_of(WHITESPACE);
 	return (start == std::string::npos) ? "" : s.substr(start);
 }
 std::string rtrim(const std::string& s)
 {
-	size_t end = s.find_last_not_of(WHITESPACE);
+	int end = s.find_last_not_of(WHITESPACE);
 	return (end == std::string::npos) ? "" : s.substr(0, end + 1);
 }
 std::string trim(const std::string& s)
